+++ dora/core/admin.py
@@ -1,0 +1,13 @@
+from django.contrib import admin
+
+
+class EnumAdmin(admin.ModelAdmin):
+    list_display = [
+        "value",
+        "label",
+    ]
+    search_fields = (
+        "value",
+        "label",
+    )
+    ordering = ["label"]
